+++ src/script/Script.ts
@@ -1,3 +1,4 @@
+import { Step } from './Step';
 import { pumlhorse } from './PumlhorseGlobal';
 import { ModuleLoader, ModuleLocator } from './ModuleLoader';
 import { ScriptInterrupt } from './ScriptInterrupt';
@@ -13,8 +14,6 @@ import { ModuleRepository } from './Modules';
 import validateScriptDefinition from './scriptDefinitionValidator';
 import * as loggers from './loggers';
 import * as helpers from '../util/helpers';
-import * as stringParser from './StringParser';
-import * as Expression from 'angular-expressions';
 import './modules/assert';
 import './modules/async';
 import './modules/conditional';
@@ -45,13 +44,14 @@ export class Script implements IScript {
         this.name = scriptDefinition.name;
         this.internalScript = new InternalScript(this.id);
 
+    }
+
+    async run(context?: any): Promise<any> {
         this.loadModules();
 
         this.loadFunctions();
         this.loadCleanupSteps();
-    }
-
-    async run(context?: any): Promise<any> {
+        
         const scope = new Scope(this.internalScript, context);
         
         try {
@@ -201,156 +201,4 @@ class InternalScript implements IScriptInternal {
 
         await step.run();
     }
-
-}
-
-const assignmentRegex = /([a-zA-Z0-9_-]+) = (.+)/;
-
-class Step {
-    assignment: string;
-    functionName: string;
-    private runFunc: Function;
-
-    constructor(funcName: string, 
-        private parameters: any,
-        private scope) {
-        const match = funcName.match(assignmentRegex);
-
-        if (match == null) {
-            this.assignment = null;
-            this.functionName = funcName;
-        }
-        else {
-            this.assignment = match[1];
-            this.functionName = match[2];
-        }
-    }
-
-    private isAssignment(): boolean {
-        return this.assignment != null;
-    }
-
-    async run() {
-        if (this.isAssignment() && this.assignment.length == 0) {
-            throw new Error('Assignment statement must have a variable name');
-        }
-
-        this.runFunc = helpers.objectByString<Function>(this.scope, this.functionName);
-
-        if (this.runFunc == null) {
-            if (this.parameters == null) {
-                this.doAssignment(await this.runSimpleStep());
-                return;
-            }
-
-            throw new Error(`Function "${this.functionName}" does not exist`);
-        }
-
-        await this.runComplexStep();
-    }
-
-    // Run a step that does not contain any parameters
-    private async runSimpleStep() {
-        return await doEval(this.functionName, true, this.scope);
-    }
-
-    private async runComplexStep() {
-        var evalParameters = null;
-        if (this.parameters != null) {
-            if (_.isArray(this.parameters)) evalParameters = this.parameters.map(p => this.compileParameter(p))
-            else if (_.isString(this.parameters)) evalParameters = this.compileParameter(this.parameters)
-            else if (_.isObject(this.parameters)) evalParameters = _.mapObject(this.parameters, (value, key) => this.compileParameter(value, key))
-            else evalParameters = this.compileParameter(this.parameters)
-        }
-
-        var functionParameterNames = helpers.getParameters(this.runFunc);
-
-        const aliases = StepFunction.getAliases(this.runFunc);
-        var params = await Bluebird.mapSeries(functionParameterNames, 
-                (name) => this.getParameter(evalParameters, name, aliases));
-        
-        let passedParams;
-                            
-        if (evalParameters === null) passedParams = null;
-        else if (helpers.isValueType(evalParameters)) passedParams = [evalParameters];
-        else if (_.isString(evalParameters)) passedParams = [evalParameters];
-        else if (_.isArray(evalParameters) && _.isArray(this.parameters)) passedParams = evalParameters;
-        else if (_.isString(this.parameters) && _.isObject(evalParameters)) passedParams = [evalParameters];
-        else passedParams = params.length > 0 ? params : [evalParameters];
-                
-        var result = await this.runFunc.apply(this.scope, passedParams);
-        
-        this.doAssignment(result);
-        return;
-    }
-
-    private compileParameter(value, key?) {
-        if (StepFunction.hasDeferredParameter(this.runFunc, key)) return value;
-            
-        return doEval(value, true, this.scope);
-    }
-
-    private getParameter(parameters: any, name: string, aliases: string[]): any {
-        
-        if (this.isParameterName('$scope', name, aliases)) return this.scope;        
-        if (this.isParameterName('$all', name, aliases)) return parameters;
-
-        let parameterValue = undefined;
-        if (parameters != null) {
-            parameterValue = parameters[name];
-            if (parameterValue == null && aliases != null) {
-                parameterValue = parameters[aliases[name]];
-            }
-        }
-
-        return parameterValue;
-    }
-
-    private isParameterName(expectedName: string, actualName: string, aliases: string[]): boolean {
-        return actualName == expectedName || (aliases != null && aliases[actualName] == expectedName)
-    }
-
-    private doAssignment(result: any) {
-        if (this.isAssignment()) {
-            this.scope[this.assignment] = result;
-            return result;
-        }
-        return result;
-    }
-}
-
-class StepFunction {
-    static hasDeferredParameter(func: Function, parameterName: string) : boolean {
-        if (func['__deferEval'] == null) return false;
-        return func['__deferEval'].indexOf(parameterName) > -1;
-    }
-
-    static getAliases(func: Function): string[] {
-        return func['__alias'];
-    }
-}
-
-function doEval(input: any, compile: boolean, scope: any): any {
-    if (input == null) return null;
-        
-    if (_.isString(input)) {
-        var parts = stringParser.parse(input)
-        parts = parts
-            .map((p) => {
-                return compile == true && p.isTokenized
-                    ? Expression.compile(p.value.trim())(scope)
-                    : p.value
-            })
-        return parts.length > 1 
-            ? parts.join("")
-            : parts[0]
-    }
-    
-    if (typeof input == "object") {
-        return (input.constructor === Array)
-            ? input.map(function (val) { return doEval(val, true, scope) })
-            : _.mapObject(input, (val) => doEval(val, true, scope));
-    }
-    
-    return input;
 }
