+++ src/script/modules/async.ts
@@ -3,18 +3,18 @@ import enforce from '../../util/enforce';
 import * as _ from 'underscore';
 import { pumlhorse } from '../PumlhorseGlobal';
 
-export class AsyncModule {
-    static runParallel(steps, scope: IScope) {        
+function runParallel() {        
+        var steps = Array.prototype.slice.call(arguments);
         enforce(steps, 'steps')
             .isNotNull()
             .isArray();
 
+        var scope = this;
         return Promise.all(steps.map(step => {
             var newSteps = _.flatten([step]);
             return scope.$runSteps(newSteps, scope);
         }))
     }
-}
 
 pumlhorse.module('async')
-    .function('parallel', ['$all', '$scope', AsyncModule.runParallel])
+    .function('parallel', runParallel)
