+++ src/util/enforce.ts
@@ -20,7 +20,6 @@ class Enforcement {
     }
 
     isArray(): Enforcement {
-        console.log(this.value);
         if (this.value != null && !_.isArray(this.value)) this.throwError('must be an array');
         return this;
     }
