+++ tsconfig.json
@@ -2,7 +2,7 @@
     "compileOnSave": true,
     "compilerOptions": {
         "lib": [ "es2015"],
-        "allowJs": true,
+        "allowJs": false,
         "target": "es5",
         "removeComments": true,
         "moduleResolution": "node",
